% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rest_client.R
\name{HttpRestClient}
\alias{HttpRestClient}
\title{HttpRestClient class}
\description{
it makes simple atomic CRUD requests to a particular end_point
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-HttpRestClient-new}{\code{HttpRestClient$new()}}
\item \href{#method-HttpRestClient-get}{\code{HttpRestClient$get()}}
\item \href{#method-HttpRestClient-get_dataframe}{\code{HttpRestClient$get_dataframe()}}
\item \href{#method-HttpRestClient-post_dataframe}{\code{HttpRestClient$post_dataframe()}}
\item \href{#method-HttpRestClient-post_object}{\code{HttpRestClient$post_object()}}
\item \href{#method-HttpRestClient-delete}{\code{HttpRestClient$delete()}}
\item \href{#method-HttpRestClient-clone}{\code{HttpRestClient$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HttpRestClient-new"></a>}}
\if{latex}{\out{\hypertarget{method-HttpRestClient-new}{}}}
\subsection{Method \code{new()}}{
Initialise HttpRestClient client
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HttpRestClient$new(base_url, end_point)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HttpRestClient-get"></a>}}
\if{latex}{\out{\hypertarget{method-HttpRestClient-get}{}}}
\subsection{Method \code{get()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HttpRestClient$get()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HttpRestClient-get_dataframe"></a>}}
\if{latex}{\out{\hypertarget{method-HttpRestClient-get_dataframe}{}}}
\subsection{Method \code{get_dataframe()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HttpRestClient$get_dataframe(parameters = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HttpRestClient-post_dataframe"></a>}}
\if{latex}{\out{\hypertarget{method-HttpRestClient-post_dataframe}{}}}
\subsection{Method \code{post_dataframe()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HttpRestClient$post_dataframe(df)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HttpRestClient-post_object"></a>}}
\if{latex}{\out{\hypertarget{method-HttpRestClient-post_object}{}}}
\subsection{Method \code{post_object()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HttpRestClient$post_object(obj)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HttpRestClient-delete"></a>}}
\if{latex}{\out{\hypertarget{method-HttpRestClient-delete}{}}}
\subsection{Method \code{delete()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HttpRestClient$delete(id)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HttpRestClient-clone"></a>}}
\if{latex}{\out{\hypertarget{method-HttpRestClient-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HttpRestClient$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
